Wait for sleeping beauty
Adding waitsFor check in the beforeEach method. The condition I'm
waiting for here is shamelessly borrowed from bok-choy's
wait_for_ajax._is_ajax_finished() method:
https://github.com/edx/bok-choy/blob/master/bok_choy/page_object.py#L431

This seems to fix the errors I was able to generate with the
previous commit's set_timeout-induced errors, which looked very
similar to the failing master tests, so I *think* this will fix
our woes, but since I was never able to get the tests to fail
exactly the same way I can't be sure.